Add post to Mastodon functionality
Pretty much a straight copy of the tweet functions, methods, and
assumed db structure but using the Mastodon.py library. Even though
some Mastodon instances allow for lengthy posts, my preference is
to have the posts have the same brevity as the tweets, so the same
text and media are used here.

Some key differences:
- The API only requires an access token, but each user must specify their API URL (i.e. Mastodon instance).
- Media MIME types are not inferred, right now I'm assuming all media is image/png.
- Error handling is not as robust, though we can tell if an access token is invalid (401) or the API URL doesn't exist (404).
- The API response object is a dict containing *all* metadata, which probably doesn't need to be logged in full.