src\iembot\util.py pointless-statement
In the function toot_cb, line 506 there is the seems to be pointless statement mduser["screen_name"]. Accessing a key might be a way to cause an expectation if the key does not exist but there is no exception handling here. mduser is created a few lines above and is not used later or returned.